requirements.txt
django==4.2.7
djangorestframework==3.14.0
pymongo==4.5.0
requests==2.31.0
dnspython==2.4.2